requirements.txt
google-api-python-client==2.62.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
httplib2==0.20.4
python-dotenv==0.21.0
requests==2.28.1
oauth2client==4.1.3
pyTelegramBotAPI==4.7.0
aiohttp==3.8.3
psycopg2-binary==2.9.3